// src/api/generated/models/StratColumnDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CountryDto } from './CountryDto';

export type StratColumnDto = {
    stratColumnId: string;
    identifier: string;
    countries: Array<CountryDto>;
};