// src/api/generated/models/StratUnitDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type StratUnitDto = {
    stratUnitId: string;
    identifier: string;
    stratColumnId: string;
    stratUnitParentId?: string | null;
    level: number;
    stratUnitType: string;
};